Refactor code
Code reformatted and refactored.
