+++ src/main/java/com/axiomq/starwars/jwt/JwtEmailPasswordRequest.java
@@ -6,7 +6,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class JwtEmailPasswordRequest {
-
     private String email;
     private String password;
 }
