+++ src/main/java/com/axiomq/starwars/jwt/JwtSecretKey.java
@@ -2,6 +2,7 @@ package com.axiomq.starwars.jwt;
 
 import com.axiomq.starwars.security.JwtConfig;
 import io.jsonwebtoken.security.Keys;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -9,18 +10,13 @@ import org.springframework.context.annotation.Configuration;
 import javax.crypto.SecretKey;
 
 @Configuration
+@RequiredArgsConstructor
 public class JwtSecretKey {
 
     private final JwtConfig jwtConfig;
-
-    @Autowired
-    public JwtSecretKey(JwtConfig jwtConfig) {
-        this.jwtConfig = jwtConfig;
-    }
 
     @Bean
     public SecretKey getSecretKey() {
         return Keys.hmacShaKeyFor(jwtConfig.getSecretKey().getBytes());
     }
-
 }
