+++ src/main/java/com/axiomq/starwars/services/impl/UserServiceImpl.java
@@ -67,5 +67,4 @@ public class UserServiceImpl implements UserService {
                 new ObjectNotFoundException(String.format("User with email %s does not exist.", email))
         );
     }
-
 }
