+++ src/main/java/com/axiomq/starwars/web/controllers/VoteController.java
@@ -14,9 +14,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.annotation.Nullable;
 import javax.validation.Valid;
-import java.io.IOException;
+
 import java.security.Principal;
 import java.util.List;
 import java.util.stream.Collectors;
