+++ src/main/java/com/axiomq/starwars/web/dtos/film/FilmMapper.java
@@ -1,7 +1,6 @@
 package com.axiomq.starwars.web.dtos.film;
 
 import com.axiomq.starwars.entities.Film;
-import com.axiomq.starwars.services.FilmService;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 import org.mapstruct.factory.Mappers;
