+++ src/main/java/com/axiomq/starwars/web/dtos/vote/VoteResponse.java
