+++ src/test/java/com/axiomq/starwars/services/impl/FilmServiceImplTest.java
@@ -1,6 +1,7 @@
 package com.axiomq.starwars.services.impl;
 
 import com.axiomq.starwars.entities.Film;
+import com.axiomq.starwars.exceptions.ObjectNotFoundException;
 import com.axiomq.starwars.repositories.FilmRepository;
 import com.axiomq.starwars.services.FilmService;
 import org.junit.jupiter.api.BeforeEach;
@@ -9,12 +10,10 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
-import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class FilmServiceImplTest {
@@ -51,7 +50,7 @@ class FilmServiceImplTest {
 
         //when
         when(filmRepository.findById(film.getId())).thenReturn(Optional.empty());
-        Exception e = assertThrows(NoSuchElementException.class, () -> {
+        Exception e = assertThrows(ObjectNotFoundException.class, () -> {
             filmService.getFilmById(film.getId());
         });
